+++ config.py
@@ -15,4 +15,4 @@ App Config
 - You shouldn't change these settings unless you know what they do
 #################################################################"""
 
-VERSION = "3.2.2"
+VERSION = "3.2.3"